requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.50.0
faiss-cpu
pypdf
tiktoken
python-dotenv